+++ sw/airborne/firmwares/tutorial/main_demo5.c
@@ -66,7 +66,7 @@ static inline void main_dl_parse_msg(void) {
   uint8_t msg_id = IdOfMsg(dl_buffer);
   if (msg_id == DL_SETTING) {
     uint8_t i = DL_SETTING_index(dl_buffer);
-    float var = DL_SETTING_value(dl_buffer);
+    double var = DL_SETTING_value(dl_buffer);
     DlSetting(i, var);
     DOWNLINK_SEND_DL_VALUE(&i, &var);
   }
